requirements.txt
flask>=2.0.0
rembg>=2.0.0
Pillow>=9.0.0
werkzeug>=2.0.0
